Move snapcrafters organization transferration task to right after stable release
It is likely that the upstream didn't respond to the upstreaming process
or are unwilling to maintain the snap package themselves, we should
better transfer the snap to the snapcrafters organization in advance as
it is likely to succeed in the first place.  We could always transfer it
back to the upstream whenever the upstream are willing to maintain the
snap.

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>